Move CLI to separate entry point module
